Update my fork from sc master

+++ segmented_data/.scripts/create_translations.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+
 import pathlib
 import os
 import sys
@@ -11,6 +13,7 @@ import argparse
 parser = argparse.ArgumentParser()
 parser.add_argument('--source', required=True, type=str, help='Source folder, e.g.: root/pli/ms')
 parser.add_argument('--target', required=True, type=str, help='Target folder, e.g.: translation/en/sujato')
+parser.add_argument('-n', '--dry-run', action='store_true', help="Perform no actions")
 group = parser.add_mutually_exclusive_group(required=True)
 group.add_argument('--uid', default=None, type=str, help='UID of division/subdivision to translate, e.g.: dn')
 group.add_argument('--uids', default=None, type=str, help='UIDs of suttas to translate, e.g.: dn1,dn2,dn3')
@@ -55,15 +58,19 @@ for file in source_files:
     new_parent = target_dir / file.parent.relative_to(source_dir)
     new_name = file.stem.split('_')[0] + '_' + target_muids + '.json'
     new_file = new_parent / new_name 
-    print(str(new_file))
     new_files.append(new_file)
     if not new_file.exists():
-        new_parent.mkdir(parents=True, exist_ok=True)
-        with new_file.open('w') as f:
-            json.dump({}, f, indent=2)
+        if args.dry_run:
+            print(f'Create {str(new_file)}')
+        else:
+            new_parent.mkdir(parents=True, exist_ok=True)
+            with new_file.open('w') as f:
+                json.dump({}, f, indent=2)
     
-
-subprocess.run(['git', 'add', *new_files])
+if args.dry_run:
+    print(f'git add {len(new_files)} files')
+else:
+    subprocess.run(['git', 'add', *new_files])
